pyproject.toml
[project]
name = "motor-emulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "adafruit-ampy>=1.1.0",
    "esptool>=4.8.1",
    "micropython-esp32-esp32-generic-c3-stubs>=1.23.0.post2",
]
